Fix consumption recording bug and improve numeric precision
- Fixed bug where internal consumption was recorded in 'Ventas' instead of 'Consumo_interno'.
- Updated popularity calculation to include internal consumption data.
- Replaced 'parseInt' with 'parseFloat' for stock and quantity handling across POS, Consumption, and Inventory modules.
- Implemented FIFO logic improvements with 0.0001 epsilon and 4-decimal precision rounding for lot updates.
- Refactored consumption logic for better clarity.

+++ js/inventory.js
@@ -59,7 +59,7 @@ const inventory = (function() {
 
     function addItemToList() {
         const selectEl = document.getElementById('inv-product-select');
-        const quantity = parseInt(document.getElementById('inv-quantity').value);
+        const quantity = parseFloat(document.getElementById('inv-quantity').value);
         const cost = parseFloat(document.getElementById('inv-cost').value);
         const price = parseFloat(document.getElementById('inv-price').value);
         const expiry = document.getElementById('inv-expiry').value;
